# api-docs/swagger.yaml
basePath: /cosmos
definitions:
...
  configapi.UpdateRouteRequest:
    properties:
      newRoute:
        $ref: '#/definitions/utils.ProxyRouteConfig'
      operation:
        type: string
      routeName:
        type: string
    required:
    - operation
    type: object
...
  github_com_azukaar_cosmos-server_src_utils.Config:
    properties:
      APITokens:
        additionalProperties:
          $ref: '#/definitions/utils.APITokenConfig'
        type: object
      Roles:
        additionalProperties:
          $ref: '#/definitions/utils.RoleConfig'
        type: object
      adminConstellationOnly:
        type: boolean
      adminWhitelistIPs:
        items:
          type: string
        type: array
      agentMode:
        type: boolean
      autoUpdate:
        type: boolean
      backup:
        $ref: '#/definitions/utils.BackupConfig'
      backupOutputDir:
        type: string
      betaUpdates:
        type: boolean
      blockedCountries:
        items:
          type: string
        type: array
      constellationConfig:
        $ref: '#/definitions/utils.ConstellationConfig'
      countryBlacklistIsWhitelist:
        type: boolean
      cron:
        additionalProperties:
          $ref: '#/definitions/utils.CRONConfig'
        type: object
      database:
        $ref: '#/definitions/utils.DatabaseConfig'
      disableHostModeWarning:
        type: boolean
      disableOpenIDDirect:
        type: boolean
      disableUserManagement:
        type: boolean
      dockerConfig:
        $ref: '#/definitions/utils.DockerConfig'
      emailConfig:
        $ref: '#/definitions/utils.EmailConfig'
      homepageConfig:
        $ref: '#/definitions/utils.HomepageConfig'
      httpconfig:
        $ref: '#/definitions/utils.HTTPConfig'
      incrBackupOutputDir:
        type: string
      licence:
        type: string
      loggingLevel:
        type: string
      marketConfig:
        $ref: '#/definitions/utils.MarketConfig'
      mongoDB:
        type: string
      monitoringAlerts:
        additionalProperties:
          $ref: '#/definitions/utils.Alert'
        type: object
      monitoringDisabled:
        type: boolean
      mpdn_:
        type: string
      mpdu_:
        type: string
      newInstall:
        type: boolean
      openIDClients:
        items:
          $ref: '#/definitions/utils.OpenIDClient'
        type: array
      remoteStorage:
        $ref: '#/definitions/utils.RemoteStorageConfig'
      requireMFA:
        type: boolean
      serverCountry:
        type: string
      serverToken:
        type: string
      storage:
        $ref: '#/definitions/utils.StorageConfig'
      themeConfig:
        $ref: '#/definitions/utils.ThemeConfig'
    required:
    - emailConfig
    - httpconfig
    type: object
...
  main.NewInstallJSON:
    properties:
      allowHTTPLocalIPAccess:
        type: boolean
      dnsChallengeProvider:
        type: string
      dnschallengeConfig:
        additionalProperties:
          type: string
        type: object
      hostname:
        type: string
      httpsCertificateMode:
        type: string
      mongodb:
        type: string
      mongodbMode:
        type: string
      nickname:
        type: string
      omitempty:
        type: string
      password:
        type: string
      sslEmail:
        type: string
      step:
        type: string
      tlsCert:
        type: string
      tlsKey:
        type: string
      useWildcardCertificate:
        type: boolean
    type: object
  main.SetupJSON:
    properties:
      DNSChallengeConfig:
        additionalProperties:
          type: string
        type: object
      allowHTTPLocalIPAccess:
        type: boolean
      clearConfig:
        description: 'Optional: clear config before setup'
        type: boolean
      constellationConfig:
        description: 'Optional: Constellation connect'
        type: string
      createAdminToken:
        description: 'Optional: create admin API token'
        type: boolean
      dnsChallengeProvider:
        type: string
      email:
        type: string
      hostname:
        description: HTTPS
        type: string
      httpsCertificateMode:
        type: string
      licence:
        description: |-
          Optional: Pro licence key — stored on the lighthouse config and
          validated via utils.ProcessLicence at the end of setup.
        type: string
      mongodb:
        type: string
      mongodbMode:
        description: Database
        type: string
      nickname:
        description: Admin
        type: string
      password:
        type: string
      sslEmail:
        type: string
      tlsCert:
        type: string
      tlsKey:
        type: string
      useWildcardCertificate:
        type: boolean
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      status:
        example: OK
        type: string
    type: object
  utils.APITokenConfig:
    properties:
      createdAt:
        type: string
      description:
        type: string
      expiresAt:
        type: string
      ipWhitelist:
        items:
          type: string
        type: array
      name:
        type: string
      owner:
        type: string
      permissions:
        items:
          $ref: '#/definitions/utils.Permission'
        type: array
      restrictToConstellation:
        type: boolean
      tokenHash:
        type: string
      tokenSuffix:
        type: string
    type: object
...
  utils.Alert:
    properties:
      actions:
        items:
          $ref: '#/definitions/utils.AlertAction'
        type: array
      condition:
        $ref: '#/definitions/utils.AlertCondition'
      enabled:
        type: boolean
      lastTriggered:
        type: string
      name:
        type: string
      period:
        type: string
      severity:
        type: string
      throttled:
        type: boolean
      trackingMetric:
        type: string
    required:
    - name
    type: object
...
  utils.BackupConfig:
    properties:
      backups:
        additionalProperties:
          $ref: '#/definitions/utils.SingleBackupConfig'
        type: object
      disable:
        type: boolean
    type: object
  utils.CRONConfig:
    properties:
      command:
        type: string
      container:
        type: string
      crontab:
        type: string
      enabled:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
  utils.ConstellationConfig:
    properties:
      FirewallBlockedClients:
        items:
          type: string
        type: array
      NATSReplicas:
        type: integer
      constellationHostname:
        type: string
      customDNSEntries:
        items:
          $ref: '#/definitions/utils.ConstellationDNSEntry'
        type: array
      dnsadditionalBlocklists:
        items:
          type: string
        type: array
      dnsblockBlacklist:
        type: boolean
      dnsdisabled:
        type: boolean
      dnsfallback:
        type: string
      dnsport:
        type: string
      doNotSyncNodes:
        type: boolean
      enabled:
        type: boolean
      iprange:
        type: string
      overrideNebulaExitNodeInterface:
        type: string
      thisDeviceName:
        type: string
      tunnels:
        items:
          $ref: '#/definitions/utils.ProxyRouteConfig'
        type: array
    type: object
  utils.ConstellationDNSEntry:
    properties:
      key:
        type: string
      type:
        type: string
      value:
        type: string
    required:
    - key
    type: object
  utils.DatabaseConfig:
    properties:
      configVolume:
        type: string
      dbVolume:
        type: string
      hostname:
        type: string
      password:
        type: string
      puppetMode:
        type: boolean
      username:
        type: string
      version:
        type: string
    type: object
  utils.DockerConfig:
    properties:
      defaultDataPath:
        type: string
      skipPruneImages:
        type: boolean
      skipPruneNetwork:
        type: boolean
    type: object
  utils.EmailConfig:
    properties:
      allowInsecureTLS:
        type: boolean
      enabled:
        type: boolean
      from:
        type: string
      host:
        type: string
      notifyLogin:
        type: boolean
      password:
        type: string
      port:
        type: string
      useTLS:
        type: boolean
      username:
        type: string
    type: object
  utils.HTTPConfig:
    properties:
      DNSChallengeConfig:
        additionalProperties:
          type: string
        type: object
      acceptAllInsecureHostname:
        type: boolean
      allowHTTPLocalIPAccess:
        type: boolean
      allowSearchEngine:
        type: boolean
      authPrivateKey:
        type: string
      authPublicKey:
        type: string
      cacert:
        type: string
      caprivateKey:
        type: string
      disablePropagationChecks:
        type: boolean
      dnschallengePropagationWait:
        type: integer
      dnschallengeProvider:
        type: string
      dnschallengeResolvers:
        type: string
      forceHTTPSCertificateRenewal:
        type: boolean
      generateMissingAuthCert:
        type: boolean
      hostname:
        type: string
      httpport:
        maxLength: 6
        minLength: 1
        type: string
      httpscertificateMode:
        type: string
      httpsport:
        maxLength: 6
        minLength: 1
        type: string
      overrideWildcardDomains:
        type: string
      proxyConfig:
        $ref: '#/definitions/utils.ProxyConfig'
      publishMDNS:
        type: boolean
      selfTLSCert:
        type: string
      selfTLSKey:
        type: string
      selfTLSKeyHostsCached:
        items:
          type: string
        type: array
      selfTLSValidUntil:
        type: string
      sslemail:
        type: string
      tlscert:
        type: string
      tlskey:
        type: string
      tlskeyHostsCached:
        items:
          type: string
        type: array
      tlsvalidUntil:
        type: string
      trustedProxies:
        items:
          type: string
        type: array
      useForwardedFor:
        type: boolean
      useWildcardCertificate:
        type: boolean
    required:
    - hostname
    - httpport
    - httpsport
    type: object
  utils.HTTPErrorResult:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  utils.HomepageConfig:
    properties:
      background:
        type: string
      expanded:
        type: boolean
      widgets:
        items:
          type: string
        type: array
    type: object
...
  utils.MarketConfig:
    properties:
      sources:
        items:
          $ref: '#/definitions/utils.MarketSource'
        type: array
    type: object
  utils.MarketSource:
    properties:
      name:
        type: string
      url:
        type: string
    type: object
  utils.OpenIDClient:
    properties:
      id:
        type: string
      redirect:
        type: string
      secret:
        type: string
    required:
    - id
    type: object
...
  utils.ProxyConfig:
    properties:
      routes:
        items:
          $ref: '#/definitions/utils.ProxyRouteConfig'
        type: array
    type: object
  utils.ProxyRouteConfig:
    properties:
      AdditionalTargets:
        items:
          type: string
        type: array
      LBMode:
        type: string
      LBStickyMode:
        type: boolean
      acceptInsecureHTTPSTarget:
        type: boolean
      addionalFilters:
        items:
          $ref: '#/definitions/utils.AddionalFiltersConfig'
        type: array
      adminOnly:
        type: boolean
      authEnabled:
        type: boolean
      blockAPIAbuse:
        type: boolean
      blockCommonBots:
        type: boolean
      corsorigin:
        type: string
      description:
        type: string
      disableHeaderHardening:
        type: boolean
      disableLegacyHTTPHeaders:
        type: boolean
      disabled:
        type: boolean
      extraHeaders:
        additionalProperties:
          type: string
        type: object
      hideFromDashboard:
        type: boolean
      host:
        type: string
      icon:
        type: string
      maxBandwith:
        type: integer
      mode:
        type: string
      name:
        type: string
      overwriteHostHeader:
        type: string
      pathPrefix:
        type: string
      restrictToConstellation:
        type: boolean
      skipURLClean:
        type: boolean
      smartShield:
        $ref: '#/definitions/utils.SmartShieldPolicy'
      spoofHostname:
        type: boolean
      stripPathPrefix:
        type: boolean
      target:
        type: string
      throttlePerMinute:
        type: integer
      timeout:
        $ref: '#/definitions/time.Duration'
      tunnel:
        type: string
      tunneledHost:
        type: string
      useH2C:
        type: boolean
      useHost:
        type: boolean
      usePathPrefix:
        type: boolean
      whitelistInboundIPs:
        items:
          type: string
        type: array
    required:
    - name
    - target
    type: object
  utils.RemoteStorageConfig:
    properties:
      shares:
        items:
          $ref: '#/definitions/utils.LocationRemoteStorageConfig'
        type: array
    type: object
  utils.RoleConfig:
    properties:
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/utils.Permission'
        type: array
    type: object
  utils.SingleBackupConfig:
    properties:
      autoStopContainers:
        type: boolean
      crontab:
        type: string
      crontabForget:
        type: string
      name:
        type: string
      password:
        type: string
      repository:
        type: string
      retentionPolicy:
        type: string
      source:
        type: string
    required:
    - name
    - repository
    type: object
...
  utils.StorageConfig:
    properties:
      snapRAIDs:
        items:
          $ref: '#/definitions/utils.SnapRAIDConfig'
        type: array
    type: object
  utils.ThemeConfig:
    properties:
      primaryColor:
        type: string
      secondaryColor:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for Cosmos Cloud server management
  title: Cosmos Server API
  version: 0.22.13
paths:
...
  /api/config:
    get:
      description: Returns the full server configuration (sensitive fields masked
        for non-credential users)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_azukaar_cosmos-server_src_utils.Config'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Get server configuration
      tags:
      - config
    patch:
      consumes:
      - application/json
      description: Performs route operations (add, replace, delete, move_up, move_down)
        on proxy routes
      parameters:
      - description: Route update operation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/configapi.UpdateRouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Patch route configuration
      tags:
      - config
    put:
      consumes:
      - application/json
      description: Replaces the entire server configuration (masked credential fields
        are preserved from existing config)
      parameters:
      - description: Full configuration object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_azukaar_cosmos-server_src_utils.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Update server configuration
      tags:
      - config
...
  /api/newInstall:
    post:
      consumes:
      - application/json
      description: Handles multi-step server installation including database setup,
        HTTPS configuration, and admin user creation
      parameters:
      - description: Installation step data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.NewInstallJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      summary: Process new installation steps
      tags:
      - system
...
  /api/setup:
    post:
      consumes:
      - application/json
      description: 'Performs the entire initial setup in a single call: database,
        HTTPS, admin user, optional Constellation connect, and optional admin API
        token creation. Only available when the server is in NewInstall mode.'
      parameters:
      - description: Setup configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.SetupJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      summary: One-shot server provisioning
      tags:
      - system
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"